pause archiving subprocesses when space runs out

+++ buffersync.sh
@@ -3,17 +3,26 @@ VERBOSE=$1
 
 #Directory that will be encrypted and transfered
 ENCRYPTTARGET="~"
+ENCRYPTTARGET=`eval echo $ENCRYPTTARGET`
 #Directory where the splitted encrypted archives will be stored before transfer
 KEYFILE="~/.ssh/id_rsa.crt"
+KEYFILE=`eval echo $KEYFILE`
 
 #Size of parts
-SPLITSIZE="50m"
+SPLITSIZE="1m"
 PARTSDIR="~/Downloads/backuptemp"
+PARTSDIR=`eval echo $PARTSDIR`
 
 #Threshold of space left in dir to pause the archiving process (kilobytes)
-SPACELEFTTHRESHOLD="500000" #approx 500M
+SPACELEFTTHRESHOLD="512000" #approx 500M
+MINSTARTSPACE="512000"      #approx 500M
 
-ENCRYPTDIR="echo \"Encrypting directory ${ENCRYPTTARGET}\";tar --ignore-failed -cpj $ENCRYPTTARGET 2>/dev/null | openssl aes-256-cbc -kfile $KEYFILE | split -d -b $SPLITSIZE - $PARTSDIR/$(date "+%Y%m%d-%s").tar.bz2.enc.; echo \"Finished encrypting $ENCRYPTTARGET\""
+function archive(){
+	echo "encrypting directory $ENCRYPTTARGET";
+	tar --exclude '$PARTSDIR/*' --exclude '$PARTSDIR' --ignore-failed -cpj $ENCRYPTTARGET 2> /dev/null | openssl aes-256-cbc -kfile $KEYFILE | split -d -b $SPLITSIZE - $PARTSDIR/$(date "+%Y%m%d-%s").tar.bz2.enc.;
+	echo "finished encrypting $ENCRYPTTARGET"
+}
+
 
 function sync1() {
 	if [[ "$VERBOSE" == "-v" ]]; then
@@ -25,7 +34,8 @@ function sync1() {
 		NEXTLOCK1=`find "$PARTSDIR" -type f -name "*flock1" 2> /dev/null | sort | head -n 1`
 		if [[ $( echo $NEXTLOCK1 | wc -c ) -gt 0 ]]; then
 			PART1=$( echo $NEXTLOCK1 | sed -e 's/.flock.*//' )
-			echo "Syncing $PART1 to USB"
+			echo "syncing $PART1 to USB"
+			sleep 10
 			rsync -tuq --modify-window=1 $PART1 /mnt/usb && rm $NEXTLOCK1 2> /dev/null
 		fi;
 		LOCKS1COUNT=$(eval $COUNTLOCKS1)
@@ -43,7 +53,7 @@ function sync2() {
 		NEXTLOCK2=`find "$PARTSDIR" -type f -name "*flock2" 2> /dev/null | sort | head -n 1`
 		if [[ $( echo $NEXTLOCK2 | wc -c ) -gt 0 ]]; then
 			PART2=$( echo $NEXTLOCK2 | sed -e 's/.flock.*//' )
-			echo "Syncing $PART2 to OASIS"
+			echo "syncing $PART2 to OASIS"
 			rsync -azq $PART2 OASIS:/mnt/lvmdisk/backup 2> /dev/null && rm $NEXTLOCK2 2> /dev/null
 		fi;
 		LOCKS2COUNT=$(eval $COUNTLOCKS2)
@@ -61,13 +71,40 @@ function sync3() {
 		NEXTLOCK3=`find "$PARTSDIR" -type f -name "*flock3" 2> /dev/null | sort | head -n 1`
 		if [[ $( echo $NEXTLOCK3 | wc -c ) -gt 0 ]]; then
 			PART3=$( echo $NEXTLOCK3 | sed -e 's/.flock.*//' )
-			echo "Syncing $PART3 to Beest"
+			echo "syncing $PART3 to Beest"
 			rsync -avq $PART3 Beest:/volume3/RAID1/backup 2> /dev/null && rm $NEXTLOCK3 2> /dev/null
 		fi;
 		LOCKS3COUNT=$(eval $COUNTLOCKS3)
 	done;
 	rm -rf "$PARTSDIR/sync3.plock"
 }
+
+function freespacelock() {
+	#Checks if we want to pause the archiving until there is enough space available
+	SPACELEFT=$(df -k $PARTSDIR | awk 'END{print $(NF-2)}')
+	if [[ $SPACELEFT -lt $SPACELEFTTHRESHOLD ]]; then
+		if [[ "$VERBOSE" == "-v" ]]; then
+			echo "space left is below threshold $SPACELEFT/$SPACELEFTTHRESHOLD"
+		fi;
+		if [ ! -d "$PARTSDIR/archive.plock" ]; then
+			#Pauses archiving process chain (like ^z)
+			echo "pausing chain ($SPACELEFT available, need $SPACELEFTTHRESHOLD)";
+			mkdir $PARTSDIR/archive.plock 2> /dev/null;
+			for childprocess in $(ps -o pid --ppid $WAITPID | tail -n +2); do kill -TSTP $childprocess; done
+		fi;
+	else
+		if [[ "$VERBOSE" == "-v" ]]; then
+			echo "space left is adequate $SPACELEFT/$SPACELEFTTHRESHOLD"
+		fi;
+		if [ -d "$PARTSDIR/archive.plock" ]; then
+			#Continues archiving process chain (like fg)
+			echo "coninuing chain ($SPACELEFT available)";
+			rm -rf "$PARTSDIR/archive.plock" 2> /dev/null;
+			for childprocess in $(ps -o pid --ppid $WAITPID | tail -n +2); do kill -CONT $childprocess; done
+		fi;
+	fi;
+}
+
 
 #Callback for the black hole handler
 function process_file() {
@@ -76,45 +113,37 @@ function process_file() {
 
 #Callback that persists every iteration of the black hole while loop
 function process_pool() {
-	#Checks if we want to pause the archiving until there is enough space available
-	SPACELEFT=$(df -k $PARTSDIR | awk 'END{print $(NF-2)}')
-	if [[ $SPACELEFT -lt $SPACELETTHRESHOLD ]]; then
-		#Pauses archiving process chain (like ^z)
-		kill -STOP WAITPID
-	else
-		#Continues archiving process chain (like fg)
-		kill -CONT WAITPID
-	fi;
+	freespacelock
 
 	if ! mkdir $PARTSDIR/sync1.plock 2> /dev/null ; then
 		if [[ "$VERBOSE" == "-v" ]]; then
-			echo "Sync 1 in progress.."
+			echo "sync 1 in progress.."
 		fi;
 	else
 		if [[ "$VERBOSE" == "-v" ]]; then
-			echo "Starting new sync1"
+			echo "starting new sync1"
 		fi;
 		sync1 &
 	fi;
 
 	if ! mkdir $PARTSDIR/sync2.plock 2> /dev/null ; then
 		if [[ "$VERBOSE" == "-v" ]]; then
-			echo "Sync 2 in progress.."
+			echo "sync 2 in progress.."
 		fi;
 	else
 		if [[ "$VERBOSE" == "-v" ]]; then
-			echo "Starting new sync2"
+			echo "starting new sync2"
 		fi;
 		sync2 &
 	fi;
 
 	if ! mkdir $PARTSDIR/sync3.plock 2> /dev/null ; then
 		if [[ "$VERBOSE" == "-v" ]]; then
-			echo "Sync 3 in progress.."
+			echo "sync 3 in progress.."
 		fi;
 	else
 		if [[ "$VERBOSE" == "-v" ]]; then
-			echo "Starting new sync3"
+			echo "starting new sync3"
 		fi;
 		sync3 &
 	fi;
@@ -136,6 +165,9 @@ function process_pool() {
 				rm "${PROCESSED}.flock0" "$PROCESSED" 2>/dev/null
 			fi;
 			PROCESSEDCOUNT=$(eval $COUNTPROCESSED)
+			if [ -d "$PARTSDIR/archive.plock" ]; then
+				sleep 1
+			fi;
 		done;
 		if [[ "$VERBOSE" == "-v" ]]; then
 			echo "end process pool"
@@ -146,6 +178,7 @@ function process_pool() {
 
 #This function checks if there are files in the tempdir, and if so then processes them
 function blackhole_handler() {
+	freespacelock
 	TEMPDIR="$1"
 	CALLBACKFUNCINNER="$2"
 	CALLBACKFUNCOUTER="$3"
@@ -157,10 +190,13 @@ function blackhole_handler() {
 	COUNTFILES="find \"$PARTSDIR\" -type f 2> /dev/null | sort | sed -e 's/.flock.*//' | uniq -u | wc -l"
 	FILECOUNT=$(eval $COUNTFILES)
 
+
 	#Loop if there are files left to process
 	while [ $FILECOUNT -gt 0 ]; do
+		freespacelock
+
 		PATH_CURRENT=`find "$PARTSDIR" -type f 2> /dev/null | sort | sed -e 's/.flock.*//' | sort | uniq -u | head -n 1 | sed -e 's/\.\///g'`
-		if [[ $FILECOUNT -gt 1 || $CLEAR_FLIP -eq 1 ]]; then
+		if [[ $FILECOUNT -gt 1 || $CLEAR_FLIP -eq 1 ]] && [[ ! -d "$PARTSDIR/archive.plock" ]]; then
 			BYTECOUNT_PREV=0
 			CLEAR_FLIP=0
 			touch $PATH_CURRENT.flock0
@@ -169,8 +205,8 @@ function blackhole_handler() {
 			#Count the amount of bytes in the current file
 			BYTECOUNT_CURRENT=$(cat $PATH_CURRENT | wc -c)
 			if [[ "$VERBOSE" == "-v" ]]; then
-				echo "BYTECOUNT CURRENT IS $BYTECOUNT_CURRENT"
-				echo "BYTECOUNT PREV IS $BYTECOUNT_PREV"
+				echo "bytecount current is $BYTECOUNT_CURRENT"
+				echo "bytecount previous is $BYTECOUNT_PREV"
 			fi;
 
 			#Check if the current file is still being written to
@@ -186,13 +222,16 @@ function blackhole_handler() {
 
 		if ! mkdir $PARTSDIR/outer.plock 2> /dev/null ; then
 			if [[ "$VERBOSE" == "-v" ]]; then
-				echo "Sync in progress"
+				echo "sync in progress"
 			fi;
 		else
-			eval $CALLBACKFUNCOUTER &
+			(eval $CALLBACKFUNCOUTER &) #Parenthesis because running this command in a subshell will prevent the [pid] done output
 		fi;
 
 		FILECOUNT=$(eval $COUNTFILES)
+		if [ -d "$PARTSDIR/archive.plock" ]; then
+			sleep $WAITDELAY
+		fi;
 	done
 }
 
@@ -201,9 +240,12 @@ function wait_while(){
 	WAITFUNC="$1"
 	LOOPFUNC="$2"
 	WAITDELAY="$3"
-	eval $WAITFUNC &
+	archive &
 	WAITPID=$!
-	while ps -p $WAITPID > /dev/null; do 
+	while ps -p $WAITPID > /dev/null || [ -d "$PARTSDIR/archive.plock" ] ; do 
+		if [[ "$VERBOSE" == "-v" ]]; then
+			echo "archive not finished yet, continuing with the callback loop";
+		fi;
 		eval $LOOPFUNC
 		sleep $WAITDELAY
 	done
@@ -211,13 +253,17 @@ function wait_while(){
 
 PARTSDIR=`eval echo $PARTSDIR`
 if [[ "$VERBOSE" == "-v" ]]; then
-	echo "PARTSDIR IS $PARTSDIR";
+	echo "partsdir is $PARTSDIR";
 fi;
-
 mkdir -p $PARTSDIR
-rm -R $PARTSDIR/*plock 2> /dev/null
+SPACELEFT=$(df -k $PARTSDIR | awk 'END{print $(NF-2)}')
+if [[ $SPACELEFT -gt $MINSTARTSPACE ]]; then
+	rm -R $PARTSDIR/*plock 2> /dev/null
 
-#Initialize the pool handler
-SYNCBLACKHOLE="blackhole_handler $PARTSDIR process_file process_pool 10"
-wait_while "$ENCRYPTDIR" "$SYNCBLACKHOLE" 1; echo "Now waiting for synchronization pool to be emptied"; wait && rm -R $PARTSDIR
-echo "Done synchronizing"
+	#Initialize the pool handler
+	SYNCBLACKHOLE="blackhole_handler $PARTSDIR process_file process_pool 10"
+	wait_while "$ENCRYPTDIR" "$SYNCBLACKHOLE" 1; echo "now waiting for synchronization pool to be emptied"; wait && rm -R $PARTSDIR
+	echo "done synchronizing"
+else
+	echo "drive it too full to start archiving (${SPACELEFT}k left need ${MINSTARTSPACE}k)";
+fi;
